+++ tests/Main.hs
@@ -34,7 +34,7 @@ main = do
 reformat :: Config -> S.ByteString -> ByteString
 reformat cfg code =
   either (("-- " <>) . L8.pack . HIndent.prettyParseError) L.fromStrict
-    $ HIndent.reformat cfg [] Nothing code
+    $ HIndent.reformat cfg HIndent.defaultExtensions Nothing code
 
 -- | Convert the Markdone document to Spec benchmarks.
 toSpec :: [Markdone] -> Spec